+++ CHANGELOG.markdown
@@ -1,3 +1,9 @@
+0.5
+---
+* The default `Numeric.Interval` now deals more conventionally with empty intervals.
+* The old "Kaucher directed interval" behavior is available as `Numeric.Interval.Kaucher`.
+* Strictly Non-Empty intervals are now contained in `Numeric.Interval.NonEmpty`
+
 0.4.2
 -----
 * Added `clamp`